exercicio 27 reescrito codigo mudado nome da classe
